move pathschanged to separate package
