+++ internal/plugin/convert/pathschanged/pathschanged.go
@@ -1,18 +1,18 @@
-package plugin
+package pathschanged
 
 import (
 	"context"
 
 	"github.com/drone/drone-go/drone"
 	"github.com/drone/drone-go/plugin/converter"
-	pathschanged "github.com/meltwater/drone-convert-pathschanged/plugin"
+	"github.com/meltwater/drone-convert-pathschanged/plugin"
 )
 
 type PathsChanged struct {
 	provider string
 }
 
-func NewPathsChanged() *PathsChanged {
+func New() *PathsChanged {
 	return &PathsChanged{
 		provider: "github",
 	}
@@ -21,7 +21,7 @@ func NewPathsChanged() *PathsChanged {
 func (p *PathsChanged) Convert(ctx context.Context, req *converter.Request) (*drone.Config, error) {
 	// TODO: this type can be removed once the upstream pathschanged
 	// plugin supports reading the drone token directly
-	plugin := pathschanged.New(p.provider, &pathschanged.Params{Token: req.Token.Access})
+	plugin := plugin.New(p.provider, &plugin.Params{Token: req.Token.Access})
 
 	return plugin.Convert(ctx, req)
 }
