--- internal/plugin/pathschanged_test.go
@@ -1,9 +1,0 @@
-package plugin
-
-import (
-	"testing"
-)
-
-func TestPathsChangedFulfillsPluginInterface(t *testing.T) {
-	_ = NewRouter("", WithConvertPlugins(NewPathsChanged()))
-}
